// frontend/src/api/generated/models/bOMDetailItemResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface BOMDetailItemResponse {
  id: string;
  material_id: string;
  material_code: string;
  material_name: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  quantity: string;
  usage_uom_code: string;
  purchase_uom_code: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  conversion_ratio: string;
  last_purchase_price?: string | null;
}