// frontend/src/api/generated/models/bOMDetailResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */
import type { BOMDetailItemResponse } from './bOMDetailItemResponse.ts';
import type { BOMDetailVariantResponse } from './bOMDetailVariantResponse.ts';

export interface BOMDetailResponse {
  id: string;
  product_variant_id: string;
  product_variant: BOMDetailVariantResponse;
  version: number;
  effective_date: string;
  is_active: boolean;
  items: BOMDetailItemResponse[];
}